requirements.txt
aio-pika==9.5.7
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiormq==6.9.0
aiosignal==1.4.0
amqp==5.2.0
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.9.1
async-timeout==5.0.1
asyncpg>=0.29
attrs==25.3.0
backports.tarfile==1.2.0
billiard==4.2.0
celery==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorlog==6.8.2
couchbase==4.3.0
debugpy==1.8.2
distro==1.9.0
dnspython==2.6.1
docker==7.1.0
docutils==0.22
email_validator==2.2.0
exceptiongroup==1.2.1
fastapi==0.111.0
fastapi-cli==0.0.4
frozenlist==1.7.0
googleapis-common-protos==1.70.0
grpcio==1.74.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
id==1.5.0
idna==3.7
importlib_metadata==8.7.0
iniconfig==2.0.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.3.0
Jinja2==3.1.4
jiter==0.10.0
jsonschema==4.23.0
keyring==25.6.0
kombu==5.3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
more-itertools==10.8.0
multidict==6.6.4
nh3==0.3.0
numpy==2.0.0
ollama==0.5.3
openai==1.104.2
opentelemetry-api==1.36.0
opentelemetry-exporter-otlp==1.36.0
opentelemetry-exporter-otlp-proto-common==1.36.0
opentelemetry-exporter-otlp-proto-grpc==1.36.0
opentelemetry-exporter-otlp-proto-http==1.36.0
opentelemetry-instrumentation==0.57b0
opentelemetry-instrumentation-aio-pika==0.57b0
opentelemetry-instrumentation-asgi==0.57b0
opentelemetry-instrumentation-dbapi==0.57b0
opentelemetry-instrumentation-fastapi==0.57b0
opentelemetry-instrumentation-httpx==0.57b0
opentelemetry-instrumentation-logging==0.57b0
opentelemetry-instrumentation-requests==0.57b0
opentelemetry-propagator-b3==1.36.0
opentelemetry-proto==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
opentelemetry-util-http==0.57b0
orjson==3.10.6
packaging==24.1
pamqp==3.3.0
pandas==2.2.2
ping3==4.0.8
pluggy==1.5.0
prompt_toolkit==3.0.47
propcache==0.3.2
protobuf==6.32.0
ptvsd==4.3.2
py-consul==1.5.1
pyshacl==0.26.0
pydantic==2.11.7
pydantic-settings==2.5.2
pydantic_core==2.33.2
Pygments==2.18.0
pytest==8.3.2
pytest-asyncio==0.24.0
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==3.3.0
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
rdflib==7.0.0
readme_renderer==44.0
requests==2.32.3
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.7.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
tenacity==9.1.2
tomli==2.0.1
tqdm==4.67.1
twine==6.2.0
typer==0.12.3
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.2
uvicorn==0.30.1
uvloop==0.19.0
vine==5.1.0
watchfiles==0.22.0
wcwidth==0.2.13
websockets==12.0
wrapt==1.17.3
yarl==1.20.1
zipp==3.23.0
